protected all routes for security

+++ Frontend/Scripts/addProduct.js
@@ -1,10 +1,26 @@
-
-// let data = JSON.parse(localStorage.getItem("admin_data"));
-
-// document.querySelector("#admin_name").innerText = localStorage.getItem("admin_name")
-// document.querySelector("#img_nav").setAttribute("src", data.usertype);
-
 const url = `http://localhost:8080`;
+
+const token=localStorage.getItem("admin")
+window.addEventListener("load",()=>{
+    if(token){
+        display()
+    }
+    else{
+        // alert("Login first")
+    
+        Swal.fire({
+            icon: 'error',
+            title: 'Oops...',
+            text: 'SalonLex - Please Login first!',
+            footer: '<a href="../Frontend/login.html">Log in</a>'
+          }).then((res)=>{
+            window.location.href = "login.html"
+          })
+    }
+})
+
+function display(){
+
 let sidebar = document.querySelector(".sidebar");
 let sidebarBtn = document.querySelector(".sidebarBtn");
 sidebarBtn.onclick = function () {
@@ -131,16 +147,40 @@ function displayCards(data) {
         }
       }
 
-userDetails();
-function userDetails() {
-    let admin = JSON.parse(localStorage.getItem("adminData"));
-    let cont = document.getElementById("admin_name");
-    let cont2 = document.getElementById("img-admin");
-
-    cont2.innerHTML = `<img src="${admin[0].image}">`
-    cont.innerHTML = `${admin[0].name}`
-};
-
-document.getElementsByClassName("log_out")[0].addEventListener("click", () => {
-     localStorage.clear("admin");
-});
+      userDetails();
+      function userDetails() {
+          let admin = JSON.parse(localStorage.getItem("adminData"));
+          let cont = document.getElementById("admin_name");
+          let cont2 = document.getElementById("img-admin");
+      
+          cont2.innerHTML = `<img src="${admin[0].image}">`
+          cont.innerHTML = `${admin[0].name}`
+      };
+      
+      document.getElementsByClassName("log_out")[0].addEventListener("click", async(e) => {
+          e.preventDefault()
+          try{
+              let res=await fetch(`${url}/users/logout`,{
+                          headers:{
+                              "content-type":"application/json",
+                              "authorization":token
+                          }
+                       })
+                      let data= await res.json()
+                      if(data.msg=="Logout Success"){
+                           localStorage.clear("admin");
+                        //    alert("logout succes")
+                           Swal.fire(
+                            'Good job!',
+                            '<h3> Logout Successfully! See You Soon 👍</h3>',
+                            'success'
+                          ).then((res)=>{
+                            window.location.href = "index.html"
+                          })
+                      }
+          }catch(err){
+              console.log(err)
+          }
+      
+      });
+}
